changed real type to be double instead of double long

+++ src/common.h
@@ -12,7 +12,7 @@
 #define min(a,b) (((a)<(b))?(a):(b))
 #define max(a,b) (((a)>(b))?(a):(b))
 
-typedef double long real;
+typedef double real;
 
 typedef struct {
  real r;
